# Write throughput results as a PR comment