--- a/.azure-pipelines/ultimate-pipeline.yml
+++ b/.azure-pipelines/ultimate-pipeline.yml
@@ -4076,6 +4076,57 @@ stages:
     - script: tracer\build.cmd CheckSmokeTestsForErrors
       displayName: CheckSmokeTestsForErrors
 
+- stage: compare_throughput
+  condition: and(succeeded(), eq(variables.isMainRepository, true))
+  dependsOn: [throughput, throughput_appsec, master_commit_id]
+  variables:
+    masterCommitId: $[ stageDependencies.master_commit_id.fetch.outputs['set_sha.master']]
+  jobs:
+  - job: compare
+    timeoutInMinutes: 60 #default value
+    pool:
+      vmImage: windows-2022
+
+    steps:
+    - template: steps/clone-repo.yml
+      parameters:
+        masterCommitId: $(masterCommitId)
+
+    - template: steps/install-latest-dotnet-sdk.yml
+
+    - task: DownloadPipelineArtifact@2
+      displayName: Download crank_linux_arm64_1
+      inputs:
+        artifact: crank_linux_arm64_1 #only download the first lot, ignores retries
+        path: $(System.DefaultWorkingDirectory)/tracer/build_data/throughput/pr/crank_linux_arm64_1
+
+    - task: DownloadPipelineArtifact@2
+      displayName: Download crank_linux_x64_1
+      inputs:
+        artifact: crank_linux_x64_1 #only download the first lot, ignores retries
+        path: $(System.DefaultWorkingDirectory)/tracer/build_data/throughput/pr/crank_linux_x64_1
+
+    - task: DownloadPipelineArtifact@2
+      displayName: Download crank_linux_x64_asm_1
+      inputs:
+        artifact: crank_linux_x64_asm_1 #only download the first lot, ignores retries
+        path: $(System.DefaultWorkingDirectory)/tracer/build_data/throughput/pr/crank_linux_x64_asm_1
+
+    - task: DownloadPipelineArtifact@2
+      displayName: Download crank_windows_x64_1
+      inputs:
+        artifact: crank_windows_x64_1 #only download the first lot, ignores retries
+        path: $(System.DefaultWorkingDirectory)/tracer/build_data/throughput/pr/crank_windows_x64_1
+
+    - script: tracer\build.cmd CompareThroughputResults
+      continueOnError: true
+      displayName: Compare Crank results
+      condition: and(succeeded(), eq(variables.isPullRequest, true))
+      env:
+        PR_NUMBER: $(System.PullRequest.PullRequestNumber)
+        AZURE_DEVOPS_TOKEN: $(AZURE_DEVOPS_TOKEN)
+        GITHUB_TOKEN: $(GITHUB_TOKEN)
+
 - stage: trace_pipeline
   condition: eq(variables['isBenchmarksOnlyBuild'], 'False')
   dependsOn:
@@ -4118,6 +4169,7 @@ stages:
     - msi_installer_smoke_tests
     - tracer_home_smoke_tests
     - coverage
+    - compare_throughput
     - system_tests
   jobs:
   - job: build_and_run