--- a/tracer/build/_build/Build.GitHub.cs
+++ b/tracer/build/_build/Build.GitHub.cs
@@ -19,6 +19,7 @@ using Nuke.Common.Tools.Git;
 using Octokit;
 using Octokit.GraphQL;
 using Octokit.GraphQL.Model;
+using ThroughputComparison;
 using YamlDotNet.Serialization.NamingConventions;
 using static Nuke.Common.IO.CompressionTasks;
 using static Nuke.Common.IO.FileSystemTasks;
@@ -953,6 +954,121 @@ partial class Build
 
              await HideCommentsInPullRequest(prNumber, "## Benchmarks Report");
              await PostCommentToPullRequest(prNumber, markdown);
+         });
+
+    Target CompareThroughputResults => _ => _
+         .Unlisted()
+         .DependsOn(CreateRequiredDirectories)
+         .Requires(() => AzureDevopsToken)
+         .Requires(() => GitHubRepositoryName)
+         .Requires(() => GitHubToken)
+         .Executes(async () =>
+         {
+             if (!int.TryParse(Environment.GetEnvironmentVariable("PR_NUMBER"), out var prNumber))
+             {
+                 Logger.Warn("No PR_NUMBER variable found. Skipping throughput comparison");
+                 return;
+             }
+
+             var testedCommit = Environment.GetEnvironmentVariable("OriginalCommitId"); 
+             if (string.IsNullOrEmpty(testedCommit))
+             {
+                 Logger.Warn("No OriginalCommitId variable found. Skipping throughput comparison");
+                 return;
+             }
+
+             var throughputDir = BuildDataDirectory / "throughput";
+             var masterDir = throughputDir / "master";
+             var oldBenchmarksDir = throughputDir / "benchmarks_2_9_0";
+             var latestBenchmarksDir = throughputDir / "latest_benchmarks";
+             var prDir = throughputDir / "pr";
+
+             FileSystemTasks.EnsureCleanDirectory(masterDir);
+             FileSystemTasks.EnsureCleanDirectory(oldBenchmarksDir);
+             FileSystemTasks.EnsureCleanDirectory(latestBenchmarksDir);
+
+             // Connect to Azure DevOps Services
+             var connection = new VssConnection(
+                 new Uri(AzureDevopsOrganisation),
+                 new VssBasicCredential(string.Empty, AzureDevopsToken));
+
+             using var buildHttpClient = connection.GetClient<BuildHttpClient>();
+
+             // Grab the comparison artifacts
+             var masterBuild = await GetCrankArtifacts(buildHttpClient, "refs/heads/master", masterDir);
+             var oldBenchmarkBuild = await GetCrankArtifacts(buildHttpClient, "refs/heads/benchmarks/2.9.0", oldBenchmarksDir);
+             var (newBenchmarkBuild, benchmarkVersion) = await GetCrankArtifactsForLatestBenchmarkBranch(buildHttpClient, latestBenchmarksDir);
+             
+             var sources = new List<CrankResultSource>
+             {
+                 new($"This PR ({PullRequestNumber})", testedCommit, CrankSourceType.Pr, prDir),
+                 new("master", masterBuild.SourceVersion, CrankSourceType.Master, masterDir),
+                 new("benchmarks/2.9.0", oldBenchmarkBuild.SourceVersion, CrankSourceType.OldBenchmark, oldBenchmarksDir),
+             };
+
+             if (newBenchmarkBuild is not null && benchmarkVersion is not null)
+             {
+                 sources.Add(new($"benchmarks/{benchmarkVersion}", newBenchmarkBuild.SourceVersion, CrankSourceType.LatestBenchmark, latestBenchmarksDir));
+             }
+
+             var markdown = CompareThroughput.GetMarkdown(sources);
+
+             await HideCommentsInPullRequest(prNumber, "## Throughput/Crank Report");
+             await PostCommentToPullRequest(prNumber, markdown);
+
+             async Task<(Microsoft.TeamFoundation.Build.WebApi.Build, string Version)> GetCrankArtifactsForLatestBenchmarkBranch(BuildHttpClient httpClient, AbsolutePath directory)
+             {
+                 // current (not released version)
+                 var version = new Version(Version);
+                 var versionsToCheck = 3;
+                 while (versionsToCheck > 0)
+                 {
+                     // only looking back across minor releases (ignoring patch etc)
+                     versionsToCheck--;
+                     version = new Version(version.Major, version.Minor - 1, 0);
+
+                     try
+                     {
+                         var thisVersion = $"{version.Major}.{version.Minor}.0";
+                         var build = await GetCrankArtifacts(httpClient, $"refs/heads/benchmarks/{thisVersion}", directory);
+                         return (build, thisVersion);
+                     }
+                     catch (Exception)
+                     {
+                         // if this fails, it's because we have no primary artifacts for that branch
+                         Console.WriteLine($"No artifacts found for version {version}, checking next branch");
+                     }
+                 }
+
+                 Console.WriteLine("No benchmarks found, skipping");
+                 return (null, null);
+             }
+
+             async Task<Microsoft.TeamFoundation.Build.WebApi.Build> GetCrankArtifacts(BuildHttpClient httpClient, string branch, AbsolutePath directory)
+             {
+                 // find the first build with the linux crank results
+                 var (build, _) = await FindAndDownloadAzureArtifact(httpClient, branch, build => "crank_linux_x64_1", directory, buildReason: null);
+
+                 // get all the other artifacts from the same build for consistency
+                 var artifacts = new[] { "crank_linux_arm64_1", "crank_linux_x64_asm_1", "crank_windows_x64_1" };
+                 foreach (var artifactName in artifacts)
+                 {
+                     try
+                     {
+                         var artifact = await httpClient.GetArtifactAsync(
+                                        project: AzureDevopsProjectId,
+                                        buildId: build.Id,
+                                        artifactName: artifactName);
+                         await DownloadAzureArtifact(directory, artifact, AzureDevopsToken);
+                     }
+                     catch (ArtifactNotFoundException)
+                     {
+                         Console.WriteLine($"Could not find {artifactName} artifact for build {build.Id}. Skipping");
+                     }
+                 }
+
+                 return build;
+             }
          });
 
     async Task PostCommentToPullRequest(int prNumber, string markdown)