## Summary of changes

Posts a summary of the throughput tests for a PR as a comment on GitHub

## Reason for change

We want to make it easier to see regressions in performance, similar to the way we do for code coverage and benchmarks

## Implementation details

We started saving the throughput/crank results as build artifacts in #3581. After this PR, we compare the throughput results for the commit being run against the latest master results, and the latest named benchmark results. We do a simple comparison of number of requests, upload the report as an artifact, and (for PRs) write a comment on GitHub.

Currently this only compares the "total requests" field. We could compare more fields, or do some averaging, but I think this is good enough, and we can always view the dashboards on DD if we need to 

## Test coverage

Ran an individual (non PR) test, and will see the results on this one too

## Other details

The report (and PR comment) will look something like this:

---

## Throughput/Crank Report:zap:

Throughput results for AspNetCoreSimpleController comparing the following branches/commits:
- [This commit (60191c)](https://github.com/DataDog/dd-trace-dotnet/tree/60191c4cc6ec5d9955d427d820c5b2f224bc9c2c)
- [master](https://github.com/DataDog/dd-trace-dotnet/tree/a8abc21d58ea9d91bc0b5b2ba5f9d265f97e41d9)
- [benchmarks/2.9.0](https://github.com/DataDog/dd-trace-dotnet/tree/0d56b1d9f25673f5257b1b7df97a3754d2df63f4)
- [benchmarks/2.21.0](https://github.com/DataDog/dd-trace-dotnet/tree/994d914bd3a83ab18a13a54b2a0907dc1d7e6b86)

Cases where throughput results for the PR are worse than latest master (5% drop or greater), results are shown in **red**.

Note that these results are based on a _single_ point-in-time result for each branch. For full results, see [one](https://ddstaging.datadoghq.com/dashboard/fnz-afi-c2i/apm-net-crank) of the [many](https://ddstaging.datadoghq.com/dashboard/c5g-i7d-kcy/ciapp-apm-net-throughput-tests), [many](https://ddstaging.datadoghq.com/dashboard/uxh-m8j-qhi/ciapp-apm-net-throughput-tests-kevin) dashboards!

```mermaid
gantt
    title Throughput Linux x64 (Total requests) 
    dateFormat  X
    axisFormat %s
    section Baseline
    This commit (60191c) (6.243M)   : 0, 6243213
    master (6.145M)   : 0, 6144635
    benchmarks/2.21.0 (6.245M)   : 0, 6244879
    benchmarks/2.9.0 (6.201M)   : 0, 6200566

    section Instrumented
    This commit (60191c) (4.198M)   : 0, 4198472
    master (4.038M)   : 0, 4038357
    benchmarks/2.21.0 (4.279M)   : 0, 4279078
    benchmarks/2.9.0 (4.423M)   : 0, 4422687

    section Trace stats
    master (4.150M)   : 0, 4150070
    benchmarks/2.21.0 (4.133M)   : 0, 4132576

```
```mermaid
gantt
    title Throughput Linux arm64 (Total requests) 
    dateFormat  X
    axisFormat %s
    section Baseline
    This commit (60191c) (5.156M)   : 0, 5155688
    master (5.598M)   : 0, 5597511
    benchmarks/2.21.0 (5.481M)   : 0, 5481181
    benchmarks/2.9.0 (5.424M)   : 0, 5423903

    section Instrumented
    This commit (60191c) (3.654M)   : 0, 3654052
    master (3.699M)   : 0, 3699122
    benchmarks/2.21.0 (3.925M)   : 0, 3924769

    section Trace stats
    master (3.606M)   : 0, 3605547
    benchmarks/2.21.0 (3.591M)   : 0, 3591431

```
```mermaid
gantt
    title Throughput Windows x64 (Total requests) 
    dateFormat  X
    axisFormat %s
    section Baseline
    This commit (60191c) (5.041M)   : 0, 5041393
    master (5.982M)   : 0, 5981706
    benchmarks/2.21.0 (6.118M)   : 0, 6117965
    benchmarks/2.9.0 (6.197M)   : 0, 6197101

    section Instrumented
    This commit (60191c) (3.915M)   : 0, 3915024
    master (4.062M)   : 0, 4061626
    benchmarks/2.21.0 (4.078M)   : 0, 4077786
    benchmarks/2.9.0 (4.196M)   : 0, 4196298

    section Trace stats
    master (4.125M)   : 0, 4124689
    benchmarks/2.21.0 (4.094M)   : 0, 4094000

```
```mermaid
gantt
    title Throughput Linux x64 (ASM) (Total requests) 
    dateFormat  X
    axisFormat %s
    section Baseline
    This commit (60191c) (3.516M)   : 0, 3516254
    master (3.463M)   : 0, 3462667
    benchmarks/2.21.0 (3.512M)   : 0, 3512411
    benchmarks/2.9.0 (3.643M)   : 0, 3643314

    section No attack
    This commit (60191c) (1.265M)   : 0, 1265018
    master (1.261M)   : 0, 1261202
    benchmarks/2.21.0 (1.266M)   : 0, 1266269
    benchmarks/2.9.0 (1.273M)   : 0, 1272523

    section Attack
    This commit (60191c) (1.088M)   : 0, 1087768
    master (1.080M)   : 0, 1079760
    benchmarks/2.21.0 (1.094M)   : 0, 1094454
    benchmarks/2.9.0 (1.093M)   : 0, 1092875

    section Blocking
    This commit (60191c) (2.125M)   : 0, 2124784
    master (2.106M)   : 0, 2106359
    benchmarks/2.21.0 (2.081M)   : 0, 2081171

```